Refactor serviceTest
